requirements.txt
################
# Django stuff #
################
django==3.0.3
django-redis==4.11.0
django-braces==1.14.0
django-filter==2.2.0
django-environ==0.4.5
django-tables2==2.2.1
django-crispy-forms==1.9.0
django-debug-toolbar==2.2

#########
# Utils #
#########
gevent==1.4.0
python-dateutil==2.8.1
uWSGI==2.0.18


####################
# Database related #
####################
redis==2.10.6
hiredis==1.0.1
py_lru_cache==0.1.4
pymongo==3.10.1
pymysql==0.9.3
DBUtils==1.3
